2) The get extent aggregate function is not visiting all the branches 
of the RTree root node when computing the bound.

3) The SDF provider filter capability does not does not advertise that it supports the spatial inside, this code change throws an exception if the inside operator is used.
4) A dynamic cast failed on Linux and caused a Rtree based filter optimization to be skipped. This change replaces the dynamic cast with a hard cast and explicit FDO type checking. i.e it the Fdo object type is as expected, we use hard cast.
5)The SDF filter is not being validated at execution time and exception of not found property are being thrown when accessing the reader. A new validator is added and called by select,update and delete when the execute method is called. Also added a better message for the case of a class not found.


git-svn-id: http://svn.osgeo.org/fdo/trunk@2322 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/SDF/Src/Provider/SdfUpdate.cpp
@@ -1,6 +1,6 @@
 // 
 //  
-//  Copyright (C) 2004-2006  Autodesk, Inc.
+//  Copyright (C) 2004-2007  Autodesk, Inc.
 //  
 //  This library is free software; you can redistribute it and/or
 //  modify it under the terms of version 2.1 of the GNU Lesser
@@ -80,7 +80,12 @@ FdoInt32 SdfUpdate::Execute()
             
     //get feature class
     FdoPtr<FdoClassDefinition> clas = FdoPtr<FdoClassCollection>(
-        m_connection->GetSchema()->GetClasses())->GetItem(m_className->GetName());
+        m_connection->GetSchema()->GetClasses())->FindItem(m_className->GetName());
+    if( clas == NULL )
+        throw FdoException::Create(NlsMsgGetMain(FDO_NLSID(SDFPROVIDER_75_CLASS_NOTFOUND), m_className->GetName()));
+
+    if( m_filter != NULL )
+        SdfQueryOptimizer::ValidateFilter( clas, m_filter );
 
 	m_connection->FlushAll( clas, true );
